chore(lIncSubs/consolidate): Longest Increasing Subsequence
Tests and Problem consolidated.
